+++ CHANGELOG.md
@@ -6,6 +6,23 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.9.0] - 2021-11-01
+### Added
+- Disable email verification [#280](https://github.com/rokwire/core-building-block/issues/280)
+
+### Changed
+- Dissociate permissions from applications [#207](https://github.com/rokwire/core-building-block/issues/207)
+
+## [1.8.0] - 2021-10-27
+### Added
+- Prepare Core for Admin app integration [#247](https://github.com/rokwire/core-building-block/issues/247)
+- Expose does account exist services API [#255](https://github.com/rokwire/core-building-block/issues/255)
+
+### Fixed
+- MongoDB ChangeStream Watch() does not recover [#259](https://github.com/rokwire/core-building-block/issues/259)
+
+## [1.7.0] - 2021-10-25
 ### Added
 - Add password reset from client and reset links [#216](https://github.com/rokwire/core-building-block/issues/216)
 - Add authentication required auth wrapper [#250](https://github.com/rokwire/core-building-block/issues/250)
@@ -111,7 +128,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Users authentication polish [#155](https://github.com/rokwire/core-building-block/issues/155)
 - Optimise the Mongo DB collections indexes usage [#146](https://github.com/rokwire/core-building-block/issues/146)
 
-[Unreleased]: https://github.com/rokwire/core-building-block/compare/v1.6.0...HEAD
+[Unreleased]: https://github.com/rokwire/core-building-block/compare/v1.9.0...HEAD
+[1.9.0]: https://github.com/rokwire/core-building-block/compare/v1.8.0...v1.9.0
+[1.8.0]: https://github.com/rokwire/core-building-block/compare/v1.7.0...v1.8.0
+[1.7.0]: https://github.com/rokwire/core-building-block/compare/v1.6.0...v1.7.0
 [1.6.0]: https://github.com/rokwire/core-building-block/compare/v1.5.0...v1.6.0
 [1.5.0]: https://github.com/rokwire/core-building-block/compare/v1.4.0...v1.5.0
 [1.4.0]: https://github.com/rokwire/core-building-block/compare/v1.3.0...v1.4.0